backend/requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.4
Django==4.0
django-cors-headers==3.11.0
django-import-export==2.7.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
filelock==3.4.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
pip==23.0.1
pipenv==2021.11.23
platformdirs==2.4.0
pytz==2021.3
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.2
tablib==3.1.0
tzdata==2022.7
virtualenv==20.10.0
virtualenv-clone==0.5.7
xlrd==2.0.1
xlwt==1.3.0
django-jsonfield==1.4.1
numpy==1.22.1
pandas==2.0.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
django-environ==0.11.2
gunicorn==22.0.0
MarkupPy==1.14
packaging==24.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
setuptools==65.5.1
wheel==0.43.0
drf-spectacular==0.27.2
formulas==1.2.8
tabulate